Remove v1 code and regression test
