+++ src/grblogtools/__init__.py
@@ -1,4 +1,4 @@
 __version__ = "2.0.0"
 
 from grblogtools.api import get_dataframe, parse
-from grblogtools.grblogtools import plot
+from grblogtools.plotting import plot
